Add app.json
Addin an app.json for two reasons:

- Ability to (technically) add a deploy button to the readme
- Allow for Review apps

The second one is the feature here. If we setup a pipeline on Heroku
this means that a PR can be deployed to a functioning application
Heroku with the press of a button.

It uses the app.json to setup a new app, which has the required env
variables and a redis and postgres set up and running, so it basically
removes any friction to set up an app to test out the PR.