Move specific seeding action to separate function
This (tries to )makes it easier to add something to the seeds without
breaking the existing specs. At the moment the specs expect the
`build` method of ProjectFactory to set things up in a particular
state.

If we set it up in that state we can't run an invoice simulation, so I
took that block out and put it in a different method that only gets
called while seeding