Avoid running specs twice
We don't get double the coverage when running them twice, so let's
not.

This was a bit of a wild goose change, the interwebs suggested it had
something to do with multiple rspec.configure blocks, so I made sure
we only had one and then it still kept happening.

After a lot of test runs I found out that `spec:system` is also a task
defined by the rspec/rails gem and that it wasn't being cleared out by
the `Rake::Task["spec"].clear` line.