Legacy engine fix, non-deterministic behavior
On a Mac it sorted the results differently than on Linux, which meant
that we didn't get the same package result.

Therefore there is now an additional check on the frequency to match
it with the package of the formula. If the package result frequency is
nil we assume it was meant as 'monthly'