Introducing: DataTest
Apart from the name, it does some useful things. It allows us to take
a snapshot of production data and run new code against it to verify
that we have the same (predicatable) results.

It does this by adding a rake task:

      DB_NAME=<production-copy> bundle exec rake data_test:capture_and_upload

And adding a spec to verify the data:

      bundle exec rake spec:data_test

On CI this should just work, locally you'll need S3 credentials to run them.

Ensure directory exists

Rubocop it.