Introducing: ParallelDhis2
While publishing a rather large invoice, I found out that most of the
time was spent in publishing to DHIS2.

When sending 11040 values to dhis2:

Single request:    43093.83299993351 (seconds)
Parallel requests:  8223.681999836117 (seconds)

Sending 3046 values had these results:

Single request:  11977.96700010076  (seconds)
Parallel rquests: 6049.210999626666 (seconds)

So big wins all around. I've also checked this on the DHIS2 side to
confirm that the data is arriving.

      select datavalue.lastupdated from datavalue join organisationunit ON organisationunit.organisationunitid = datavalue.sourceid join dataelement ON dataelement.dataelementid = datavalue.dataelementid where dataelement.uid = 'WQbeZuIvhzS' and organisationunit.uid = 'XJkQtdGRO5Q';

And it was updating the lastupdated, it was also growing the audit
table by the expected amounts.

ParallelDhis2 tries to mimick as much of its behavior from the normal
Dhis2::Client, so that changes to that client or configuration are
automatically taken over by ParallelDhis2.

What's with the one huge file?

1. It's basically one class with two helper classes
2. It has a potential to be reused in other projects, so one file is
quicker to test out than a couple of files
3. I thought it was going to be smaller