Fix: Data Test flow on CI
The `DataTest.all_cases` block was being executed while the file was
being evaluated, this meant that it was running before the
`before(:all)` block had a chance to download the config file.

I've also taken the opportunity to clear out the config file, since it
also has sensitive data.