Introducing Flipper: Feature Flipping
By default the code will just do the old way of publishing, if
however one was inclined to do the following:

      Flipper[:use_parallel_publishing].enable(ProjectAnchor.find(15))

One would find that from then on work for ProjectAnchor with ID: 15,
would use the parallel request sending, instead of the normal flow.

If one didn't like to do this in a console, one could also visit
`/flipper` and use the web UI (if one remembered the ADMIN_PASSWORD).