Allow a user to select org units before simulation
Some of our simulations loaded so many org units, which loaded so many
tables that the view rendering took to long and ran over the Heroku
request. To solve this we're doing two things:

- Make the view rendering async (coming to another PR)
- Allow user to select which org units to simulate (this PR)

Few things to note:

- Copying URL, will include the selected org units
- If left blank all org units will be returned
- Views have been slightly tweaked to make it clear to user.

Things I'm not sure about, the location of the `OrgUnitLimiter`
source. On the one hand, I like that it's only in the Controller. On
the other hand I'm not that much a fan of stowing it in the
controller.

It's somewhere between Presenter, ViewModel and Service.