Upgrade Rails Admin and remove font-awesome-sass
Icons were not rendering anymore, that's because the
`font-awesome-rails` gem also defines a stylesheet called
'font-awesome' and when we do a `@import 'font-awesome'` that one will
be picked, so we're back to font-awesome version 4.

I've tried:

- Removing `font-awesome-rails` from the asset paths (sass could still
find it)
- Providing an absolute path to `@import`, this worked but has the
downside of only working on my machine.
- Searching for alternatives (the best suggestion was to import all
the css and JS yourself and putting them in vendor and then use them
both this way)

In the end I settled on:

- Make it easy to upgrade to font-awesome-sass as soon as
font-awesome-rails goes away. You can remove our `icon` helper and
just use theirs and it should all just work.