ruby-2.5.1 -> ruby-2.5.8
Mainly nothing to see here:

      https://github.com/ruby/ruby/compare/v2_5_1...v2_5_8

Mainly security and bug fixes.

I tested the performance by running the test suite on this
laptop (2,8 GHz Dual-Core Intel Core i5 MacBook)

In 2.5.1: Finished in 4 minutes 55.6 seconds (files took 12.32 seconds
to load)
In 2.5.8: Finished in 4 minutes 55 seconds (files took 14.79 seconds to load)

Since I didn't run this a couple of times in a row, it's not much of a
benchmark. But I think the performance is roughly the same.