Fix: Not all response have a data_set_complete
I don't know why, but sometimes the error popups:

      https://sentry.io/organizations/bluesquareorg/issues/1850775543/?project=213257&query=is%3Aunresolved

This will make it so that it doesn't error, and maybe shows what is
the root cause of the error.

It does have responses (otherwise we would get an empty hash back in
the hash) so it needs to be having responses returned that don't carry
the `data_set_complete`, so maybe it's a one-off DHIS thing. It
doesn't happen always, but on a set of 500-ish invoices, it's bound to
happen on one of them.