Use cookies when parallel publishing
I tried to reuse the existing cookiejar we have in our DHIS gem, but
since that stores the cookies in a ruby hash, like so:

      {"JSESSIONID"	=> "393D14-and-so-on"}

And typhoeus just passes the file where the cookies are on to Ecto,
which is a wrapper for libcurl, which expect the cookies to be in a
file formatted like this:

      # Netscape HTTP Cookie File
      # https://curl.haxx.se/docs/http-cookies.html
      # This file was generated by libcurl! Edit at your own risk.

      #HttpOnly_sandbox.bluesquare.org	FALSE	/	FALSE	0	JSESSIONID	393D14-and-so-on

I tried to fake it by writing the file myself, and giving Typhoeus a
StringIO, but it really needs a file, and since it also has the domain
stuff in it, I figured this might break sooner than that it will help
us.

So, now I just use a tempfile, and the first request of Typhoeus will
fill it, and the others will use it.

So in our DHIS-logs this means one AuthenticationEvent instead of one
per publish action. Which is already a bit of a win in itself.