Change the prepared statement limit
New approach! What could happen is that when multiple projects are
generating invoices at the same time, we generate a lot of prepared
statements, which causes the OOM.

I checked the work_mem of our pg:

      select * from pg_settings where name='work_mem';
      32768kB

We have sidekiq running with a concurrency of 20, the default prepared
statement limit of Rails is 1000, so that could quickly become 20000
prepared statements that we keep in cache, and thus we might overload
our pg and get the dreaded OOM message.

This would also explain that when running them at another time, we
don't run into it.

Questions remaining:

- Is this really the way to configure a DB on Heroku? (Looks like it)
- Do we also want to run it like this on the web-dyno's? I think we
  can do a `Sidekiq.server?` like method to isolate the behavior.
- I matched the default Rails configuration but want to lower it to
  something like 200. Which might make it slower but should make it
  more dependable. It's hard to tell without more logs from Heroku.

Let's not break CI

Make test env use passed in keys or defaults

And let's pass in the host

And just to be sure the port as well

Move all DB pool size to database.yml

I've also added a small rake task to output the right number, normally
the SIDEKIQ_CONCURRENCY will always be higher than the amount of web
workers, so no need to set the DB_POOL, but just in case.

noop. Offer pearls to the linter gods

noop. Remove the copy/pasted dot